pyproject.toml
[tool.poetry]
name = "ancost"
version = "0.1.0"
description = "Test task"
authors = ["Alexander <alexpunder@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
tortoise-orm = "^0.21.6"
uvicorn = "^0.30.6"
isort = "^5.13.2"
pydantic-settings = "^2.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
